Move plot to its own component
