+++ interface/packages/react-app/package.json
@@ -38,6 +38,7 @@
     "graphql": "^14.6.0",
     "ipfs-http-client": "^45.0.0",
     "plotly.js": "^1.56.0",
+    "prop-types": "^15.7.2",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-plotly.js": "^2.5.0",
@@ -45,9 +46,6 @@
     "react-scripts": "3.4.1",
     "styled-components": "^5.1.1",
     "web3modal": "^1.9.0"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "scripts": {
     "build": "react-scripts build",
@@ -58,7 +56,6 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.11.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-config-wesbos": "^1.0.1",
