+++ interface/packages/react-app/src/components/WalletButton/WalletButton.jsx
@@ -1,19 +1,28 @@
+import { Web3Provider } from '@ethersproject/providers';
+import PropTypes from 'prop-types';
 import React from 'react';
 import { Button } from '..';
 import { logoutOfWeb3Modal } from '../../utils/web3Modal';
 
-export default function WalletButton({ provider, loadWeb3Modal }) {
+function WalletButton({ web3Provider, loadWeb3Modal }) {
   return (
     <Button
       onClick={() => {
-        if (!provider) {
+        if (!web3Provider) {
           loadWeb3Modal();
         } else {
           logoutOfWeb3Modal();
         }
       }}
     >
-      {!provider ? 'Connect Wallet' : 'Disconnect Wallet'}
+      {!web3Provider ? 'Connect Wallet' : 'Disconnect Wallet'}
     </Button>
   );
 }
+
+WalletButton.propTypes = {
+  web3Provider: PropTypes.instanceOf(Web3Provider).isRequired,
+  loadWeb3Modal: PropTypes.func.isRequired,
+};
+
+export default WalletButton;
